refactor: create file reporter and use different headers per file type
