+++ src/runner.js
@@ -1,6 +1,8 @@
 const {
-  getPackageInfoList, writeReportFile, extractInvalidPackages, formatForbiddenLicenseError
+  getPackageInfoList, extractInvalidPackages, formatForbiddenLicenseError
 } = require('./utils');
+
+const reporter = require('./reporter');
 
 const run = async (checker, args) => {
   const packages = await checker.parsePackages(args.start);
@@ -9,14 +11,14 @@ const run = async (checker, args) => {
 
   const forbiddenPackages = extractInvalidPackages(args.failOn, packageList);
   if (forbiddenPackages.length) {
-    writeReportFile(args.errorReportFileName, forbiddenPackages, args.customHeader);
+    reporter.writeErrorReportFile(args.errorReportFileName, forbiddenPackages);
     throw new Error(formatForbiddenLicenseError(forbiddenPackages));
   }
 
   const packagesIncludeLicenses = packageList.some(p => args.generateOutputOn.includes(p.licenses));
   if (!args.generateOutputOn.length || packagesIncludeLicenses) {
     if (!args.disableReport) {
-      writeReportFile(args.outputFileName, packageList, args.customHeader);
+      reporter.writeReportFile(args.outputFileName, packageList, args.customHeader);
     }
   }
 };
