+++ src/utils.js
@@ -1,8 +1,3 @@
-const fs = require('fs');
-const _ = require('lodash');
-
-const defaultReportHeader = 'This application makes use of the following open source packages:';
-
 /**
  * Generate objects with information on each package that we want to include
  * in the report files. This will include metadata such as the package name,
@@ -25,56 +20,6 @@ const getPackageInfoList = packages => Object.entries(packages)
 
     return validInfo;
   });
-
-/**
- *  Generates a table in markdown format with information on the installed
- * packages, including their licenses.
- *
- * @param {object[]} packageList - List of packages to be dumped inside of the file
- */
-const buildPackageTable = packageList => {
-  const tableHeader = '\n| Library | Version | License | Repository |\n|---|---|---|---|\n';
-
-  const compiled = _.template(
-    `${tableHeader}<% _.forEach(report, elem => { const r = /(.*)@(.*)/.exec(elem.package); %>| <%- r[1] %> | <%- r[2] %> | <%- elem.licenses %> | <%- elem.repository %> |\n<% }); %>`
-  );
-
-  return compiled({ report: packageList });
-};
-
-/**
- * Generate report file.
- *
- * @param {string} outputFileName - Name of the file for the generated report
- * @param {object[]} packageList - List of packages to be dumped inside of the file
- * @param {string} customHeaderFileName - Name of the file that contains the custom header
- *  to add at the beginning of the generated report
- */
-const writeReportFile = (outputFileName, packageList, customHeaderFileName) => {
-  let licenseReportHeader = defaultReportHeader;
-  if (customHeaderFileName) {
-    try {
-      licenseReportHeader = fs.readFileSync(customHeaderFileName);
-    } catch {
-      console.warn(`Failed to read file ${customHeaderFileName}, so default header will be added to the report`);
-    }
-  }
-
-  const licenseTable = buildPackageTable(packageList);
-
-  fs.writeFileSync(
-    `${outputFileName}.md`,
-    `${licenseReportHeader}\n${licenseTable}`,
-    error => {
-      if (error) {
-        console.error(`Error generating report file ${outputFileName}.md`);
-        throw error;
-      }
-    }
-  );
-
-  console.info(`${outputFileName}.md created!`);
-};
 
 /**
  * Parses failOn arguments distinguishing between plain string and
@@ -139,7 +84,6 @@ const formatForbiddenLicenseError = licenses => {
 
 module.exports = {
   getPackageInfoList,
-  writeReportFile,
   extractInvalidPackages,
   parseFailOnArgs,
   formatForbiddenLicenseError
